remove shared.utils and utils.general, moved a few edited functions back into utils.utils and rerouted imports back to utils.utils
